feat: add provenance attestation
This adds a new `--provenance` flag to npm for provenance attestation
during `npm publish`

If set to `true`, npm will detect if it is running in GitHub actions and
will generate an appropriate attestation bundle for that environment.

If given a path, npm will use the specified attestation and attach it
directly to the published package.

The primary work in this PR was done by
[@bdehamer](https://github.com/bdehamer), with some cleanup and
edge-case handling added by the npm cli team.